Move elf parsing into a separate library
